Support Numba GPU arrays as a backend for tests
This command works with Open MPI 4.0.1 (compiled with the --with-cuda flag):

    python test/runtests.py --numba --no-thread -v

The implementation is a minimum requirement for circumventing many Numba bugs,
which is unfortunate. Even so, we still need Numba 0.46 + the two patches from
numba/numba#4629 and numba/numba#5012 to make it fully work.

TODO: figure out the minimum supported Numba version and error out when the
installed version is lower than that.